requirements.txt
fastapi
uvicorn
langchain
langchain-community
google-generativeai
python-dotenv
sentence-transformers
chromadb
